requirements.txt
PyYAML==5.1.1
PyMySQL==0.9.3
urllib3==1.25.3